feat(tests): Add comprehensive test coverage and security tests
- Add LLM and Cognee integration verification tests
- Add AI debate maximal challenge test suite
- Add penetration test suite for LLM security
- Add unit tests for routing/semantic, testing/llm packages
- Add unit tests for agentic, memory, observability packages
- Fix RAG integration tests and type mismatches
- Add monitoring reports for test runs

Test coverage improvements:
- internal/routing/semantic: 96.2% coverage
- internal/testing/llm: 96.2% coverage
- All 10 LLM providers verified with mock tests
- Cognee integration verified
- Security tests for prompt injection, jailbreaking, data exfiltration

+++ internal/rag/hyde_test.go
@@ -369,7 +369,7 @@ func TestHyDEGenerator_HyDESearch(t *testing.T) {
 
 	assert.Equal(t, "test query", result.OriginalQuery)
 	assert.Len(t, result.Results, 2)
-	assert.Equal(t, float32(0.9), result.Results[0].Score)
+	assert.Equal(t, 0.9, result.Results[0].Score)
 	assert.Equal(t, 2, result.TopK)
 }
